+++ CoreUtilities/test/unit/src/au/gov/asd/tac/constellation/utilities/gui/filechooser/ShowFileChooserDialogNGTest.java
@@ -16,12 +16,13 @@
 package au.gov.asd.tac.constellation.utilities.gui.filechooser;
 
 import java.io.File;
+import java.util.List;
+import java.util.Optional;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import org.openide.filesystems.FileChooserBuilder;
-import static org.testng.Assert.assertNull;
-import static org.testng.Assert.assertSame;
+import static org.testng.Assert.assertEquals;
 import org.testng.annotations.Test;
 
 /**
@@ -40,11 +41,11 @@ public class ShowFileChooserDialogNGTest {
         final ShowFileChooserDialog showFileChooserDialog = new ShowFileChooserDialog(
                 fileChooserBuilder, FileChooserMode.SAVE);
         
-        assertNull(showFileChooserDialog.getSelectedFile());
+        assertEquals(showFileChooserDialog.getSelectedFiles(), Optional.empty());
         
         showFileChooserDialog.run();
         
-        assertSame(showFileChooserDialog.getSelectedFile(), file);
+        assertEquals(showFileChooserDialog.getSelectedFiles().get(), List.of(file));
         
         verify(fileChooserBuilder).showSaveDialog();
     }
@@ -59,12 +60,32 @@ public class ShowFileChooserDialogNGTest {
         final ShowFileChooserDialog showFileChooserDialog = new ShowFileChooserDialog(
                 fileChooserBuilder, FileChooserMode.OPEN);
         
-        assertNull(showFileChooserDialog.getSelectedFile());
+        assertEquals(showFileChooserDialog.getSelectedFiles(), Optional.empty());
         
         showFileChooserDialog.run();
         
-        assertSame(showFileChooserDialog.getSelectedFile(), file);
+        assertEquals(showFileChooserDialog.getSelectedFiles().get(), List.of(file));
         
         verify(fileChooserBuilder).showOpenDialog();
+    }
+    
+    @Test
+    public void run_open_multi() {
+        final FileChooserBuilder fileChooserBuilder = mock(FileChooserBuilder.class);
+        final File file1 = mock(File.class);
+        final File file2 = mock(File.class);
+        
+        when(fileChooserBuilder.showMultiOpenDialog()).thenReturn(new File[] {file1, file2});
+        
+        final ShowFileChooserDialog showFileChooserDialog = new ShowFileChooserDialog(
+                fileChooserBuilder, FileChooserMode.MULTI);
+        
+        assertEquals(showFileChooserDialog.getSelectedFiles(), Optional.empty());
+        
+        showFileChooserDialog.run();
+        
+        assertEquals(showFileChooserDialog.getSelectedFiles().get(), List.of(file1, file2));
+        
+        verify(fileChooserBuilder).showMultiOpenDialog();
     }
 }